Refactor email handling in EmailController and Email helper; add reset password functionality with new templates and remove user verification email logic from UserController.

+++ Controllers/EmailController.js
@@ -1,18 +1,22 @@
-const { createEmailAndSend, OTPTemplate, generateOTP } = require("../Helper/Email.js");
-const client = require('../Helper/Redis.js')
+const { createEmailAndSend, changePasswordTemplate, generateOTP, resetPasswordTemplate } = require("../Helper/Email.js");
+const client = require('../Helper/Redis.js');
+const User = require("../Models/User.js");
 
 // API to send OTP
 const sendEmail = {
     validator: async (req, res, next) => {
-        const { email } = req.body;
+        const { email, isResetPassword } = req.body;
         if (!email) return res.status(400).json({ message: "Email is required" });
         next();
     },
     controller: async (req, res) => {
-        const { email } = req.body;
+        const { email, isResetPassword=false } = req.body;
+        const user = await User.findOne({email, isDeleted: false, isActive: true}).lean();
+        if (!user) return res.status(400).json({ message: "User not found or not active , please check your email" });
+
         const otp = generateOTP();
-        const emailTemplate = OTPTemplate(otp);
-        const subject = "Your OTP Code";
+        const emailTemplate = isResetPassword ? resetPasswordTemplate(user.name, user.lastName, otp) : changePasswordTemplate(user.name, user.lastName, otp);
+        const subject = isResetPassword ? "Reset Password OTP" : "Password Change Request";
         try {
             await client.del(email);
             await createEmailAndSend(email, subject, emailTemplate, otp);
